Split Range into main/detail headers and unify Pointers
